requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.0.5
django-extensions==3.0.2
django-import-export==2.5.0
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
filelock==3.0.12
graphviz==0.14
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.4
incremental==21.3.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.2
mysql-connector-python==8.0.20
odfpy==1.4.1
openpyxl==3.0.7
platformdirs==2.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.4.1
service-identity==21.1.0
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
virtualenv==20.7.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
zope.interface==5.4.0
